# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_volunteer_application_approve_volunteer_applications__application_id__approval_post": {
        "properties": {
          "accepted_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted Group Id"
          }
        },
        "title": "Body_volunteer_application_approve_volunteer_applications__application_id__approval_post",
        "type": "object"
      },
      "Body_volunteer_application_group_approve_volunteer_applications_groups__group_id__approval_post": {
        "properties": {
          "accepted_group_id": {
            "title": "Accepted Group Id",
            "type": "string"
          }
        },
        "required": [
          "accepted_group_id"
        ],
        "title": "Body_volunteer_application_group_approve_volunteer_applications_groups__group_id__approval_post",
        "type": "object"
      },
      "Body_volunteer_application_mark_trial_attendance_volunteer_applications__application_id__trial_attendance_post": {
        "properties": {
          "attended": {
            "title": "Attended",
            "type": "string"
          }
        },
        "required": [
          "attended"
        ],
        "title": "Body_volunteer_application_mark_trial_attendance_volunteer_applications__application_id__trial_attendance_post",
        "type": "object"
      },
      "Body_volunteer_application_submit_apply__token__post": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "gender": {
            "default": "A",
            "title": "Gender",
            "type": "string"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "profile_photo": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Photo"
          }
        },
        "required": [
          "last_name",
          "phone"
        ],
        "title": "Body_volunteer_application_submit_apply__token__post",
        "type": "object"
      },
      "Body_volunteer_applications_create_invite_volunteer_applications_post": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "role_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Id"
          }
        },
        "required": [
          "email"
        ],
        "title": "Body_volunteer_applications_create_invite_volunteer_applications_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kvarteret Personal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/apply/{token}": {
      "get": {
        "operationId": "volunteer_application_form_apply__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Form"
      },
      "post": {
        "operationId": "volunteer_application_submit_apply__token__post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_volunteer_application_submit_apply__token__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Submit"
      }
    },
    "/apply/{token}/submitted": {
      "get": {
        "operationId": "volunteer_application_submitted_apply__token__submitted_get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Submitted"
      }
    },
...
    "/volunteer-applications": {
      "get": {
        "operationId": "volunteer_applications_index_volunteer_applications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Volunteer Applications Index"
      },
      "post": {
        "operationId": "volunteer_applications_create_invite_volunteer_applications_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_volunteer_applications_create_invite_volunteer_applications_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Applications Create Invite"
      }
    },
...
    "/volunteer-applications/groups/{group_id}/approval": {
      "post": {
        "operationId": "volunteer_application_group_approve_volunteer_applications_groups__group_id__approval_post",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_volunteer_application_group_approve_volunteer_applications_groups__group_id__approval_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Group Approve"
      }
    },
...
    "/volunteer-applications/{application_id}": {
      "delete": {
        "operationId": "volunteer_application_delete_volunteer_applications__application_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Delete"
      },
      "get": {
        "operationId": "volunteer_application_detail_volunteer_applications__application_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Detail"
      }
    },
    "/volunteer-applications/{application_id}/approval": {
      "post": {
        "operationId": "volunteer_application_approve_volunteer_applications__application_id__approval_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_volunteer_application_approve_volunteer_applications__application_id__approval_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Approve"
      }
    },
    "/volunteer-applications/{application_id}/drop-from-group": {
      "post": {
        "operationId": "volunteer_application_drop_from_group_volunteer_applications__application_id__drop_from_group_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Drop From Group"
      }
    },
    "/volunteer-applications/{application_id}/resend": {
      "post": {
        "operationId": "volunteer_application_resend_volunteer_applications__application_id__resend_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Resend"
      }
    },
    "/volunteer-applications/{application_id}/trial-attendance": {
      "post": {
        "operationId": "volunteer_application_mark_trial_attendance_volunteer_applications__application_id__trial_attendance_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_volunteer_application_mark_trial_attendance_volunteer_applications__application_id__trial_attendance_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteer Application Mark Trial Attendance"
      }
    },
    "/volunteers": {
      "get": {
        "operationId": "volunteers_index_volunteers_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "only_active",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Only Active"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Volunteers Index"
      }
    },
...
  }
}